+++ src/features/game/state/GameStateConstants.ts
