+++ src/features/game/utils/ErrorUtils.ts
