+++ src/features/game/utils/GameAccess.ts
@@ -1,3 +1,4 @@
+// to be replaced with account info
 export function hasDevAccess() {
   return true;
 }
